Makefiles refactoring
* move generic implementation of scrypt_core to scrypt-generic.c
* add USE_ASM option
