+++ src/scrypt-arm.S
@@ -8,7 +8,7 @@
  */
 
 
-#if defined(OPTIMIZED_SALSA) &&  defined(__arm__) && defined(__APCS_32__)
+#if defined(__arm__) && defined(__APCS_32__)
 
 #if defined(__linux__) && defined(__ELF__)
 .section .note.GNU-stack,"",%progbits
@@ -447,9 +447,9 @@ _scrypt_core:
 	sub	sp, sp, #21*4
 	bic	sp, sp, #63
 	str	r12, [sp, #20*4]
-	
+
 	scrypt_shuffle()
-	
+
 	str	r0, [sp, #16*4]
 	add	r12, r1, #1024*32*4
 	str	r12, [sp, #18*4]
@@ -462,9 +462,9 @@ scrypt_core_loop1:
 	scrypt_core_macro1a_x4()
 	scrypt_core_macro1a_x4()
 	str	r1, [sp, #17*4]
-	
+
 	salsa8_core()
-	
+
 	ldr	r0, [sp, #16*4]
 	mov	r12, sp
 	add	r2, r0, #16*4
@@ -472,9 +472,9 @@ scrypt_core_loop1:
 	scrypt_core_macro2_x4()
 	scrypt_core_macro2_x4()
 	scrypt_core_macro2_x4()
-	
+
 	salsa8_core()
-	
+
 	ldr	r0, [sp, #16*4]
 	mov	r1, sp
 	add	r0, r0, #16*4
@@ -483,12 +483,12 @@ scrypt_core_loop1:
 	ldr	r3, [sp, #17*4]
 	ldr	r12, [sp, #18*4]
 	scrypt_core_macro3_x4()
-	
+
 	add	r1, r3, #16*4
 	sub	r0, r0, #32*4
 	cmp	r1, r12
 	bne	scrypt_core_loop1
-	
+
 	ldr	r4, [r0, #16*4]
 	sub	r1, r1, #1024*32*4
 	str	r1, [sp, #17*4]
@@ -508,9 +508,9 @@ scrypt_core_loop2:
 	scrypt_core_macro1b_x4()
 	scrypt_core_macro1b_x4()
 	scrypt_core_macro1b_x4()
-	
+
 	salsa8_core()
-	
+
 	ldr	r0, [sp, #16*4]
 	mov	r12, sp
 	add	r2, r0, #16*4
@@ -518,9 +518,9 @@ scrypt_core_loop2:
 	scrypt_core_macro2_x4()
 	scrypt_core_macro2_x4()
 	scrypt_core_macro2_x4()
-	
+
 	salsa8_core()
-	
+
 	ldr	r0, [sp, #16*4]
 	mov	r1, sp
 	ldr	r3, [sp, #17*4]
@@ -535,15 +535,15 @@ scrypt_core_loop2:
 #endif
 	scrypt_core_macro3_x6()
 	scrypt_core_macro3_x6()
-	
+
 	ldr	r12, [sp, #18*4]
 	sub	r0, r0, #32*4
 	ldr	r1, [sp, #19*4]
 	subs	r12, r12, #1
 	bne	scrypt_core_loop2
-	
+
 	scrypt_shuffle()
-	
+
 	ldr	sp, [sp, #20*4]
 #ifdef __thumb__
 	ldmfd	sp!, {r4-r11, lr}
